+++ admin/sql/install.mysql.utf8.sql
@@ -1,5 +1,7 @@
 DROP TABLE IF EXISTS `#__lupo_agecategories`;
 DROP TABLE IF EXISTS `#__lupo_categories`;
+DROP TABLE IF EXISTS `#__lupo_clients`;
+DROP TABLE IF EXISTS `#__lupo_clients_borrowed`;
 DROP TABLE IF EXISTS `#__lupo_genres`;
 DROP TABLE IF EXISTS `#__lupo_game`;
 DROP TABLE IF EXISTS `#__lupo_game_editions`;
@@ -14,10 +16,11 @@ CREATE TABLE `#__lupo_agecategories` (
   `title` varchar(255) NOT NULL,
   `alias` varchar(255) NOT NULL DEFAULT '',
   `description` varchar(5120) NOT NULL DEFAULT '',
+  `samples` varchar(255) NOT NULL DEFAULT '',
   `published` tinyint(1) NOT NULL DEFAULT '1',
   `sort` smallint(6) DEFAULT NULL,
   PRIMARY KEY (`id`)
-) ENGINE=MyISAM DEFAULT CHARSET=utf8;
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 
 CREATE TABLE `#__lupo_categories` (
@@ -25,10 +28,35 @@ CREATE TABLE `#__lupo_categories` (
   `title` varchar(255) NOT NULL,
   `alias` varchar(255) NOT NULL DEFAULT '',
   `description` varchar(5120) NOT NULL DEFAULT '',
+  `samples` varchar(255) NOT NULL DEFAULT '',
   `published` tinyint(1) NOT NULL DEFAULT '1',
   `sort` smallint(11) DEFAULT NULL,
   PRIMARY KEY (`id`)
-) ENGINE=MyISAM DEFAULT CHARSET=utf8;
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;
+
+
+CREATE TABLE `#__lupo_clients` (
+  `adrnr` int(10) unsigned NOT NULL,
+  `username` char(20) DEFAULT NULL,
+  `firstname` varchar(50) DEFAULT NULL,
+  `lastname` varchar(50) DEFAULT NULL,
+  `aboenddat` date DEFAULT NULL,
+  `abotype` varchar(40) DEFAULT NULL,
+  UNIQUE KEY `adrnr` (`adrnr`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;
+
+
+CREATE TABLE `#__lupo_clients_borrowed` (
+  `lupo_id` int(10) unsigned NOT NULL,
+  `edition_id` int(11) DEFAULT NULL,
+  `adrnr` int(10) unsigned DEFAULT NULL,
+  `tax_extended` float DEFAULT '0',
+  `return_date` date DEFAULT NULL,
+  `return_date_extended` date DEFAULT NULL,
+  `return_extended` tinyint(4) DEFAULT NULL,
+  `return_extended_online` tinyint(4) DEFAULT '0',
+  `reminder_sent` tinyint(4) DEFAULT '0'
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 
 CREATE TABLE `#__lupo_game` (
@@ -38,16 +66,33 @@ CREATE TABLE `#__lupo_game` (
   `number` char(10) DEFAULT NULL,
   `title` char(50) DEFAULT NULL,
   `description` text,
+  `description_title` varchar(255) DEFAULT NULL,
   `fabricator` char(50) DEFAULT NULL,
+  `author` varchar(50) DEFAULT NULL,
+  `artist` varchar(50) DEFAULT NULL,
   `days` tinyint(4) DEFAULT NULL,
   `play_duration` char(35) DEFAULT NULL,
   `players` char(30) DEFAULT NULL,
   `keywords` varchar(255) DEFAULT '',
   `genres` varchar(255) DEFAULT '',
+  `prolongable` tinyint(4) DEFAULT '1',
   PRIMARY KEY (`id`),
   KEY `game_catid` (`catid`),
-  KEY `game_age_catid` (`age_catid`)
-) ENGINE=MyISAM DEFAULT CHARSET=utf8;
+  KEY `game_age_catid` (`age_catid`),
+  KEY `game_number` (`number`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;
+
+
+CREATE TABLE `#__lupo_game_documents` (
+  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `gameid` int(11) NOT NULL,
+  `code` varchar(20) NOT NULL,
+  `type` varchar(20) NOT NULL,
+  `desc` varchar(255) DEFAULT NULL,
+  `value` varchar(255) DEFAULT NULL,
+  PRIMARY KEY (`id`),
+  KEY `documents_gameid` (`gameid`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 
 CREATE TABLE `#__lupo_game_editions` (
@@ -59,26 +104,6 @@ CREATE TABLE `#__lupo_game_editions` (
   `tax` double DEFAULT NULL,
   PRIMARY KEY (`id`),
   KEY `editions_gameid` (`gameid`)
-) ENGINE=MyISAM DEFAULT CHARSET=utf8;
-
-
-CREATE TABLE `#__lupo_game_documents` (
-  `id` int NOT NULL AUTO_INCREMENT,
-  `gameid` int(11) NOT NULL,
-  `code` varchar(20) NOT NULL,
-  `type` varchar(20) NOT NULL,
-  `desc` varchar(255) DEFAULT NULL,
-  `value` varchar(255) DEFAULT NULL,
-  PRIMARY KEY (`id`),
-  KEY `documents_gameid` (`gameid`)
-)ENGINE=MyISAM DEFAULT CHARSET=utf8;
-
-
-CREATE TABLE `#__lupo_genres` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `genre` char(30) DEFAULT NULL,
-  PRIMARY KEY (`id`),
-  UNIQUE KEY `genre` (`genre`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 
@@ -95,5 +120,13 @@ CREATE TABLE `#__lupo_game_related` (
   `gameid` int(11) DEFAULT NULL,
   `number` double DEFAULT NULL,
   PRIMARY KEY (`id`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;
+
+
+CREATE TABLE `#__lupo_genres` (
+  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `genre` char(30) DEFAULT NULL,
+  PRIMARY KEY (`id`),
+  UNIQUE KEY `genre` (`genre`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;