+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### 3.4.0 (May 30, 2017)
+  - added content plugin to show nubmer of total toys 
+  - update installer SQL, change engine to InnoDB
+  - author/artist added to toy
+  - Fix php notices
+
 ### 3.3.0 (March 06, 2017)
   - show delete photos button in admin to all users
   - Fix php warning